+++ sites/geohub/src/routes/api/licenses/+server.ts
@@ -1,12 +1,4 @@
-import { type RequestHandler } from '@sveltejs/kit';
-import type { License } from '$lib/types';
-import { db } from '$lib/server/db';
-import { asc } from 'drizzle-orm';
-import { licenseInGeohub } from '$lib/server/schema';
+import api from '$api';
+import type { RequestEvent, RequestHandler } from './$types';
 
-export const GET: RequestHandler = async () => {
-	const licenses: License[] = await db.query.licenseInGeohub.findMany({
-		orderBy: [asc(licenseInGeohub.name)]
-	});
-	return new Response(JSON.stringify(licenses));
-};
+export const GET: RequestHandler = (event: RequestEvent) => api.handle(event);